Do not coerce previous release tag to semver.
When `2.0.0-beta.1` is coerced it becomes `2.0.0`. This means
a `release:major` label will release `3.0.0` instead of `3.0.0`.
Instead of `coerce` the input tag this commit uses the `clean`
command instead and errors if an invalid tag is found.

Note a minor/patch release on a major beta will release that major.
This is the default behaviour of the `semver` library. We may want
to disallow this to prevent accidental major releases.

This commit is chunky (sorry) as it also includes tests which is
run along with linting in a new Github Action.

Fixes: https://github.com/Financial-Times/origami-version/issues/56