Bundle the action's code into a single file for use
This commit stops including the node_modules directory in version control. Instead we now create a bundle using `ncc` which is what the package.json's `main` property points to.

Rather than manually updating the bundle in the codebase on every commit, I've opted to using the "tag and release" github action to update the bundle just before it adds the new tag. This removes the need for use to remember to update the bundle ourselves and brings this project's release workflow closer to our standard workflow.