throw an error if multiple release labels were applied to the merged pull-request
The current behavior which exists in origami-version is to choose the highest release label if multiple were applied. We should change this behaviour to throw an error because if multiple release labels are applied, we don't actually know that they wanted to use the highest label for the release, it is safer to throw an error.

Future work we could do in origami-version is to run the action on every pushed commit to the pull-request and fail the build if the pull-request has mulitple release labels applied. This would stop the pull-request from ever being able to be merged with multiple release labels applied.