Add ability to release beta versions
This change adds the ability for a release label named 'release:beta' to
be used to release a prerelease version.

If the current release is itself a prerelease version then the
prerelease number is incremented.
If the current version is not a prerelease then the current version has
`beta.0` appended to it.