Add ability to automatically release a major/minor/patch beta
Given the last release was not a prerelease and:
- release:major and release:beta labels are applied: do an npm version premajor
- release:minor and release:beta labels are applied: do an npm version preminor
- release:patch and release:beta labels are applied: do an npm version prepatch
- release:beta label alone is applied: do an npm version prerelease

Given the last release was a prerelease and:
- release:beta with any other release label: do an npm version
prerelease to increment the current beta.

Existing prereleases are only incremented with the `release:beta`
and any other release label to defend against users accidentally
releasing a beta for a future version which is not being worked on.
E.g. Given a PR based on a current beta `v2.0.0-beta.0`, the release
tags 'release:major' and 'release:beta' will release `v2.0.0-beta.1`,
not `v3.0.0-beta.0`.

Fixes: https://github.com/Financial-Times/origami-version/issues/114